# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Aquarium Controller API
  description: |
    REST API untuk Smart Aquarium Controller dengan fitur Automatic Feeder dan UV Sterilizer.

    ## Fitur Utama
    - **Automatic Feeder**: Sistem pemberian pakan dengan mekanisme Double Gate
    - **UV Sterilizer**: Kontrol lampu UV dengan penjadwalan dan manual override
    - **Scheduler**: Penjadwalan otomatis untuk feeding dan UV
    - **History**: Log semua aktivitas device

    ## Demo Mode
    API ini mendukung demo mode yang mensimulasikan device responses tanpa memerlukan MQTT broker atau hardware.
  version: 1.0.0
  contact:
    name: API Support
    email: support@aquarium.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://iot-backend-cursor-production.up.railway.app/api/v1
    description: Local development server (default)
  - url: https://iot-backend-cursor-production.up.railway.app/api/v1
    description: Production server
...
paths:
...
  /uv/schedules:
    get:
      tags:
        - UV
      summary: Get all UV schedules
      description: Mengambil semua jadwal UV yang telah dikonfigurasi dengan pagination
      operationId: getUVSchedules
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page (max 100)
      responses:
        "200":
          description: List jadwal UV dengan pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/UVSchedule"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags:
        - UV
      summary: Create new UV schedule
      description: Membuat jadwal UV baru
      operationId: createUVSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UVScheduleInput"
            example:
              day_name: "Monday"
              start_time: "20:00"
              end_time: "04:00"
              is_active: true
      responses:
        "201":
          description: Jadwal berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UVSchedule"
  /uv/schedules/{id}:
    put:
      tags:
        - UV
      summary: Update UV schedule
      description: Update jadwal UV yang sudah ada
      operationId: updateUVSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID jadwal UV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UVScheduleInput"
      responses:
        "200":
          description: Jadwal berhasil diupdate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UVSchedule"
        "404":
          description: Jadwal tidak ditemukan
    delete:
      tags:
        - UV
      summary: Delete UV schedule
      description: Menghapus jadwal UV
      operationId: deleteUVSchedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID jadwal UV
      responses:
        "200":
          description: Jadwal berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Schedule deleted"
  /uv/manual:
    post:
      tags:
        - UV
      summary: Trigger manual UV
      description: |
        Menyalakan UV secara manual dengan durasi tertentu.
        Manual UV memiliki prioritas override terhadap jadwal otomatis.
      operationId: manualUV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - duration_minutes
              properties:
                duration_minutes:
                  type: integer
                  minimum: 1
                  description: Durasi UV menyala dalam menit
                  example: 120
            example:
              duration_minutes: 120
      responses:
        "200":
          description: Command berhasil dikirim
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "UV command sent"
                  action_id:
                    type: integer
                    example: 15
                  duration_sec:
                    type: integer
                    example: 7200
                  end_time:
                    type: string
                    format: date-time
                    example: "2025-11-19T11:26:00Z"
        "400":
          description: Bad request (durasi tidak valid)
  /uv/manual/stop:
    post:
      tags:
        - UV
      summary: Stop running manual UV
      description: Menghentikan manual UV override sebelum durasi selesai.
      operationId: stopManualUV
      responses:
        "200":
          description: Manual UV berhasil dihentikan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Manual UV stopped"
                  action_id:
                    type: integer
                    example: 15
                  stopped_at:
                    type: string
                    format: date-time
                    example: "2025-11-19T10:00:00Z"
        "404":
          description: Tidak ada manual UV yang sedang berjalan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    UVSchedule:
      type: object
      properties:
        id:
          type: integer
          example: 1
        day_name:
          type: string
          enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun]
          example: "Mon"
        start_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          description: Waktu mulai dalam format HH:MM
          example: "20:00"
        end_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
          description: Waktu selesai dalam format HH:MM
          example: "04:00"
        is_active:
          type: boolean
          default: true
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UVScheduleInput:
      type: object
      required:
        - day_name
        - start_time
        - end_time
      properties:
        day_name:
          type: string
          enum: [Mon, Tue, Wed, Thu, Fri, Sat, Sun]
        start_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
        end_time:
          type: string
          pattern: "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
        is_active:
          type: boolean
          default: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message here"
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        page_size:
          type: integer
          description: Number of items per page
          example: 20
        total:
          type: integer
          description: Total number of records
          example: 156
        total_pages:
          type: integer
          description: Total number of pages
          example: 8